littleboss/backend/requirements.txt
boto3>=1.28.0
google-generativeai>=0.7.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
